added override for link command to insert resourcelink instead of link syntax
